Rename token file, factor out test token into separate file. Fix spelling.
