+++ Mobile Buy SDK/Mobile Buy SDK Tests/BUYIntegrationTest.m
@@ -993,8 +993,7 @@
 	[self fetchShippingRates];
 	[self updateCheckout];
 	
-	id<BUYPaymentToken> token = [self addCreditCardToCheckout];
-	[self completeCheckoutWithToken:token];
+	[self completeCheckoutWithToken:[self addCreditCardToCheckout]];
 	
 	[self pollUntilCheckoutIsComplete];
 	[self verifyCompletedCheckout];
