+++ Mobile Buy SDK/Mobile Buy SDK/Models/BUYCreditCardToken.m
@@ -41,7 +41,7 @@
 	return self;
 }
 
-#pragma mark - BUYPaymentSessionProvider -
+#pragma mark - BUYPaymentToken -
 
 - (NSDictionary *)JSONDictionary
 {
