+++ Mobile Buy SDK/Mobile Buy SDK/Static Framework/Buy.h
@@ -64,7 +64,7 @@
 #import "BUYObserver.h"
 #import "BUYShopifyErrorCodes.h"
 
-#import "BUYPaymentSessionProvider.h"
+#import "BUYPaymentToken.h"
 #import "BUYPaymentButton.h"
 #import "BUYProductViewController.h"
 #import "BUYStoreViewController.h"
